# pagos_api/src/main/java/com/poli/arq/software/pagos_api/controller/PagosController.java
package com.poli.arq.software.pagos_api.controller;

import com.poli.arq.software.pagos_api.model.Pago;
import com.poli.arq.software.pagos_api.repository.PagoRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;
import org.springframework.http.ResponseEntity;
import java.util.List;

@RestController
@RequestMapping("/api/pagos")
public class PagosController {
    @Autowired
    private PagoRepository pagoRepository;

    @GetMapping
    public ResponseEntity<List<Pago>> listarPagos() {
        return ResponseEntity.ok(pagoRepository.findAll());
    }

    @GetMapping("/{id}")
    public ResponseEntity<Pago> obtenerPago(@PathVariable Integer id) {
        return pagoRepository.findById(id)
                .map(ResponseEntity::ok)
                .orElse(ResponseEntity.notFound().build());
    }

    @PostMapping
    public ResponseEntity<Pago> registrarPago(@RequestBody Pago pago) {
        return ResponseEntity.ok(pagoRepository.save(pago));
    }
}